Extract CityGML projection object parser
